Remove the old main file & replace it with the new one :)
